4776 - Clean up user store (actions, reducers, state)
