+++ src/client/components/PageLayout/Header/UserLinks/LinksFRA/LinksFRA.tsx
@@ -15,7 +15,8 @@ import { useAppDispatch } from 'client/store/hooks'
 import { useAssessment } from 'client/store/meta/hooks/assessments'
 import { useCycle } from 'client/store/meta/hooks/cycles'
 import { AppDispatch } from 'client/store/types'
-import { UserActions, useUser } from 'client/store/user'
+import { useUser } from 'client/store/user'
+import { UserActions } from 'client/store/user/actions'
 import { useCountryIso, useIsLoginRoute } from 'client/hooks'
 import { ToasterHook, useToaster } from 'client/hooks/useToaster'
 import Icon from 'client/components/Icon'
