+++ src/client/store/user/actions/index.ts
@@ -1,8 +1,7 @@
-import { createAsyncThunk } from '@reduxjs/toolkit'
-import axios from 'axios'
+import { logout } from 'client/store/user/actions/logout'
+import { userSlice } from 'client/store/user/slice'
 
-import { ApiEndPoint } from 'meta/api/endpoint'
-
-export const logout = createAsyncThunk('user/post/logout', async () => {
-  await axios.post(ApiEndPoint.Auth.logout())
-})
+export const UserActions = {
+  ...userSlice.actions,
+  logout,
+}
