+++ src/client/store/user/index.ts
@@ -9,5 +9,3 @@ export {
   useUserCountries,
   useUserCycles,
 } from './hooks'
-export { UserActions } from './slice'
-export type { UserState } from './stateType'
