+++ src/client/store/user/slice.ts
@@ -1,41 +1,24 @@
 import { createSlice, Reducer } from '@reduxjs/toolkit'
 
-import { ApplicationActions } from 'client/store/application/actions'
-import { LoginActions } from 'client/store/login/actions'
-import { UserManagementActions } from 'client/store/ui/userManagement'
+import { getApplicationReducer } from 'client/store/user/extraReducers/getApplicationReducer'
+import { getLoginReducer } from 'client/store/user/extraReducers/getLoginReducer'
+import { getLogoutReducer } from 'client/store/user/extraReducers/getLogoutReducer'
+import { getUserManagementReducer } from 'client/store/user/extraReducers/getUserManagementReducer'
 
-import { logout } from './actions'
 import { UserState } from './stateType'
 
-const initialState: UserState = null
+export const initialState: UserState = null
 
 export const userSlice = createSlice({
   name: 'user',
   initialState,
   reducers: {},
   extraReducers: (builder) => {
-    builder.addCase(logout.fulfilled, () => initialState)
-
-    builder.addCase(ApplicationActions.initApp.fulfilled, (_, { payload }) => payload.user)
-    builder.addCase(LoginActions.localLogin.fulfilled, (_, { payload }) => payload)
-
-    builder.addCase(UserManagementActions.updateUser.fulfilled, (state, { payload }) =>
-      payload.user.id === state.id ? { ...payload.user, roles: state.roles } : state
-    )
-
-    builder.addCase(UserManagementActions.updateRoleProps.fulfilled, (state, { meta }) => {
-      const { role } = meta.arg
-      const i = state.roles.findIndex((r) => r.id === role.id)
-      if (i !== -1) state.roles[i] = { ...state.roles[i], ...role }
-    })
-
-    builder.addCase(LoginActions.acceptInvitation.fulfilled, (_, { payload }) => payload)
+    getLogoutReducer(builder)
+    getApplicationReducer(builder)
+    getUserManagementReducer(builder)
+    getLoginReducer(builder)
   },
 })
-
-export const UserActions = {
-  ...userSlice.actions,
-  logout,
-}
 
 export default userSlice.reducer as Reducer<UserState>
